automagic installation of xtb binaries on pip-install
This should finally get us to the goal of having `pip install xtbf`
actually work to make the install process much more convenient.

One idea to also make this work for non-linux systems:
Try to install it using conda first, if it is found.
That should yield the most reliable installation process in any
case. So the logic should be something like:
if xtb binary already on PATH:
   do nothing
else:
- try conda install xtb
- only if fails:
   attempt installation using urlretrieve of file